% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseGroupers.R
\name{allLeadersSpanningLeader}
\alias{allLeadersSpanningLeader}
\title{Make a leader that spans the max leader per transcript by cage files}
\usage{
allLeadersSpanningLeader(leadersFolder, dataFolder)
}
\description{
Make a leader that spans the max leader per transcript by cage files
}
