% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{artificial.from.cds}
\alias{artificial.from.cds}
\title{Artificial ORFs from CDS}
\usage{
artificial.from.cds(
  cds,
  size = 100,
  startCodons = c("ATG", "CTG", "TTG", "AAG", "AGG"),
  stopCodons = c("TAA", "TGA", "TAG"),
  fa,
  mode = "aCDS"
)
}
\arguments{
\item{cds}{GRangesList of CDS}

\item{size}{integer, default: 100, 1/6 of maximum size of ORFs (max size 600 if 100)}

\item{startCodons, }{character vector, default: c("ATG", "CTG", "TTG", "AAG", "AGG")}

\item{stopCodons}{character vector, default: c("TAA", "TGA", "TAG")}

\item{fa}{a FaFile or path to fasta index file}

\item{mode}{character, default: "aCDS", must be either CDS or aCDS (artificial CDS)}
}
\value{
GRangesList of new ORFs
}
\description{
A wrapper to ORFik:::artificial.orfs, that create artificial ORFs from CDS,
with given start and stop codons of size size * 6
}
