% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueTables.R
\name{atlasTissue}
\alias{atlasTissue}
\title{Group feature table by tissue
1st table is filtered on fpkm > 1 per tissue
2nd table is row-mean fpkm per tissue}
\usage{
atlasTissue(
  df.rfp,
  table = readTable("fpkmRFP"),
  dbOutputNames = c("RiboByTissueTF", "RiboByTissueMean")
)
}
\arguments{
\item{df.rfp}{an ORFik experiment}

\item{table}{"Ribofpkm"}

\item{dbOutputNames}{the 2 output names c("RiboByTissueTF", "RiboByTissueMean")}
}
\description{
Group feature table by tissue
1st table is filtered on fpkm > 1 per tissue
2nd table is row-mean fpkm per tissue
}
