% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineParts.R
\name{createCatalogueDB}
\alias{createCatalogueDB}
\title{Main function to fill uORF database}
\usage{
createCatalogueDB(
  df.cage,
  dataBaseFolder = get("dataBaseFolder", envir = .GlobalEnv),
  idFolder = get("idFolder", envir = .GlobalEnv),
  dataFolder = get("dataFolder", envir = .GlobalEnv),
  leadersFolder = get("leadersFolder", envir = .GlobalEnv)
)
}
\arguments{
\item{df.cage}{ORFik experiment of CAGE, set to NULL if you don't have CAGE.}
}
\description{
Step 4 of uORFome pipeline
The rows of tables will always be uORFs in order as the candidate uORF file
For transcripts it will be the transcript order in cageTx
}
