% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseCreator.R
\name{createGRObjects}
\alias{createGRObjects}
\title{convert to gr from string and filter NB!!! put this  in pipeline!!}
\usage{
createGRObjects(dataFolder, leadersFolder)
}
\description{
convert to gr from string and filter NB!!! put this  in pipeline!!
}
