% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseCreator.R
\name{createUniqueIDs}
\alias{createUniqueIDs}
\title{Create 1 column of all unique ids from uorfID folder}
\usage{
createUniqueIDs(idFolder)
}
\description{
Create 1 column of all unique ids from uorfID folder
}
