% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseCreator.R
\name{createUORFAtlas}
\alias{createUORFAtlas}
\title{Create cage ATLAS of uORFs, per CAGE sample, does uORF exist in leaders defined
by that CAGE}
\usage{
createUORFAtlas(idFolder = idFolder, dataFolder = dataFolder)
}
\description{
Will create a matrix (1 column per CAGE sample) and save to database
}
