% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseHelpers.R
\name{deleteTable}
\alias{deleteTable}
\title{Delete single sql table}
\usage{
deleteTable(tableName, uorfDB = get("uorfDB", envir = .GlobalEnv))
}
\arguments{
\item{tableName}{name of table in sql database}

\item{uorfDB}{the defined uORFome sql database, default:
get("uorfDB", envir = .GlobalEnv)}
}
\description{
Delete single sql table
}
