% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseHelpers.R
\name{deleteUorfTables}
\alias{deleteUorfTables}
\title{Delete uorf tables}
\usage{
deleteUorfTables(onlyRibo = FALSE, onlySeq = FALSE)
}
\arguments{
\item{onlyRibo}{(FALSE), only ribo-seq / RNA seq features}

\item{onlySeq}{(FALSE), only sequence features}
}
\description{
For rerunning
It keeps the RNA seq tables, since they usually not change
}
