% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseValidation.R
\name{feature.boxplots}
\alias{feature.boxplots}
\title{Box plot of feature difference between active and inactive uORFs}
\usage{
feature.boxplots(
  prediction = predictUorfs(mode = mode),
  tissue = readTable("experiment_groups")[[1]][1]
)
}
\arguments{
\item{prediction}{data.table of prediction values in c(0, 1).}

\item{tissue, }{character, default readTable("experiment_groups")[[1]][1].}
}
\value{
ggplot object of boxplots
}
\description{
Box plot of feature difference between active and inactive uORFs
}
