% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierStats.R
\name{featureAnalysis}
\alias{featureAnalysis}
\title{General feature analysis}
\usage{
featureAnalysis(prediction, tissue)
}
\arguments{
\item{prediction}{a data.table, 3 columns: (prediction, p0 and p1).
Made from uORFomePipe}
}
\description{
General feature analysis
}
