% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierStats.R
\name{findClassificationBoundary}
\alias{findClassificationBoundary}
\title{Classification bounaries}
\usage{
findClassificationBoundary(tissues)
}
\arguments{
\item{tissues}{the tissues wanted to check}
}
\description{
Classification bounaries
}
