% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{fixNAandINFTable}
\alias{fixNAandINFTable}
\title{Remove infinite and NAs in matrix}
\usage{
fixNAandINFTable(predicate)
}
\arguments{
\item{predicate}{a matrix}
}
\description{
Remove infinite and NAs in matrix
}
