% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierHelpers.R
\name{forest}
\alias{forest}
\title{Train h2o rf model.
negDT if you want own samples for that}
\usage{
forest(dt, cv = 10, ntrees = 64, nthreads = 40, max_mem_size = "200G")
}
\description{
Train h2o rf model.
negDT if you want own samples for that
}
