% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_analysis.R
\name{geneSymbolsTo}
\alias{geneSymbolsTo}
\title{Get egene symbols from Ensembl gene names}
\usage{
geneSymbolsTo(geneNames, org.db = org.Dr.eg.db)
}
\arguments{
\item{geneNames}{character vector of gene names}
}
\description{
Get egene symbols from Ensembl gene names
}
