% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicGetters.R
\name{getAll}
\alias{getAll}
\title{Get all annotation parts
leader, cds, threeUTRs and tx}
\usage{
getAll(include.cage = TRUE, cdsOnFiveEnd = FALSE)
}
\arguments{
\item{include.cage}{logical TRUE}

\item{cdsOnFiveEnd}{logical FALSE}
}
\description{
Get all annotation parts
leader, cds, threeUTRs and tx
}
