% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_analysis.R
\name{getAllORFGeneSymbols}
\alias{getAllORFGeneSymbols}
\title{Get gene symbols from ensemble gene names}
\usage{
getAllORFGeneSymbols(geneNames, dataset)
}
\arguments{
\item{geneNames}{a character vector}

\item{dataset}{default human: hsapiens_gene_ensembl,
for zebrafish drerio_gene_ensembl, for yeast scerevisiae_gene_ensembl}
}
\value{
a data.table of geneNames and symbols (2 columns)
}
\description{
Get gene symbols from ensemble gene names
}
