% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierTrainingData.R
\name{getAllSequenceFeaturesTable}
\alias{getAllSequenceFeaturesTable}
\title{Get all sequence features for ORFs}
\usage{
getAllSequenceFeaturesTable()
}
\value{
data.table with sequence features
}
\description{
Extracted from the data base
}
