% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierHelpers.R
\name{getAllUORFData}
\alias{getAllUORFData}
\title{Get all data for combined dataset}
\usage{
getAllUORFData(
  prediction = readTable("finalPredWithProb"),
  stringsAsFactors = TRUE
)
}
\description{
Get all data for combined dataset
}
