% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_analysis.R
\name{getBiomartFromOrganism}
\alias{getBiomartFromOrganism}
\title{Guess biomart from organism name}
\usage{
getBiomartFromOrganism(organism, biomart = "ensembl")
}
\arguments{
\item{organism}{scientific name of organism,
like Homo sapiens, Danio rerio, etc.}

\item{biomart}{default "ensembl", get gene symbols and GO terms for uORF genes. Will
be automaticly detected by organism name in ensembl database.
Set to NULL if you don't want to check Gene symbols and GO terms.}
}
\value{
a character with dataset used
}
\description{
Guess biomart from organism name
}
