% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicGetters.R
\name{getCageTx}
\alias{getCageTx}
\title{New leaders defined as maximum upstream for each original leader}
\usage{
getCageTx()
}
\description{
With these leader all found uORFs are guaranteed to overlap within
}
