% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineParts.R
\name{getCandidateuORFs}
\alias{getCandidateuORFs}
\title{Find uORFs from new leader regions}
\usage{
getCandidateuORFs(
  folder = regionUORFsFolder,
  startCodons = "ATG|CTG|TTG|GTG|AAG|AGG|ACG|ATC|ATA|ATT",
  stopCodons = "TAA|TAG|TGA",
  BPPARAM = bpparam()
)
}
\arguments{
\item{folder}{folder to save, default .GlobalEnv arugment: regionUORFsFolder}

\item{startCodons}{default "ATG|CTG|TTG|GTG|AAG|AGG|ACG|ATC|ATA|ATT", set to
"ATG|CTG|TTG|GTG" for a more certain set.}

\item{stopCodons}{default "TAA|TAG|TGA"}
}
\description{
Step 2 of uORFome pipeline
}
