% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicGetters.R
\name{getCDS}
\alias{getCDS}
\title{Get the coding sequences from the gtf file}
\usage{
getCDS(assignIt = TRUE, dataFolder = get("dataFolder", envir = .GlobalEnv))
}
\arguments{
\item{assignIt}{TRUE (assign to .GlobalEnv)}

\item{dataFolder}{default: get("dataFolder", envir = .GlobalEnv)}
}
\description{
Get the coding sequences from the gtf file
}
