% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicGetters.R
\name{getCDSFiltered}
\alias{getCDSFiltered}
\title{Get CDS that were filtered}
\usage{
getCDSFiltered(dataFolder = get("dataFolder", envir = .GlobalEnv))
}
\arguments{
\item{dataFolder}{default: get("dataFolder", envir = .GlobalEnv)}
}
\value{
GRangesList of all CDS used
}
\description{
With valid width modulus length, start and stop codons used
}
