% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicGetters.R
\name{getCDSTraining}
\alias{getCDSTraining}
\title{Get CDS used for positive training set}
\usage{
getCDSTraining(
  mode = "uORF",
  dataFolder = get("dataFolder", envir = .GlobalEnv)
)
}
\arguments{
\item{mode}{"uORF", or "aCDS"}

\item{dataFolder}{default: get("dataFolder", envir = .GlobalEnv)}
}
\value{
GRangesList of CDS used for training
}
\description{
Get CDS used for positive training set
}
