% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicGetters.R
\name{getFasta}
\alias{getFasta}
\title{Get the fasta indexed file}
\usage{
getFasta(filePath = NULL, assignIt = TRUE)
}
\description{
if assignIt is TRUE, the object is not return to local scope
Only assigned to globalenvir
}
