% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{getGeneralRiboFeatures}
\alias{getGeneralRiboFeatures}
\title{Get Ribo-seq features}
\usage{
getGeneralRiboFeatures(
  df.rfp,
  df.rna = NULL,
  grl,
  preName = "",
  threeUTRsSpecial = NULL
)
}
\arguments{
\item{df.rfp}{ORFik experiment}

\item{df.rna}{NULL}

\item{grl}{the ORFs as GRangesList}

\item{preName}{name to pre append in database for each feature}

\item{threeUTRsSpecial}{default NULL, or a GRangesList of 3' UTRs}
}
\value{
NULL (features saved to database)
}
\description{
Excluding features that uses RNA-seq normalizations if df.rna is NULL
}
