% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineParts.R
\name{getIDsFromUorfs}
\alias{getIDsFromUorfs}
\title{Find unique uORF ID's from uORFs}
\usage{
getIDsFromUorfs(folder = uorfFolder, BPPARAM = bpparam())
}
\arguments{
\item{folder}{folder to save, default .GlobalEnv arugment: uorfFolder}

\item{BPPARAM}{
    An instance of a \code{BiocParallelParam} class, e.g.,
    \code{\link[BiocParallel]{MulticoreParam}}, \code{\link[BiocParallel]{SnowParam}},
    \code{\link[BiocParallel]{DoparParam}}.
  }
}
\description{
Step 3 of uORFome pipeline
}
