% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineParts.R
\name{getLeadersFromCage}
\alias{getLeadersFromCage}
\title{Reassign leaders by CAGE}
\usage{
getLeadersFromCage(cageFiles, filterValue = 3, BPPARAM = bpparam())
}
\arguments{
\item{cageFiles}{a ORFik experiment with CAGE files and annotation}

\item{filterValue}{The minimum number of reads on cage position, for it to
be counted as possible new tss. (represented in score column in
CageSeq data) If you already filtered, set it to 0.}

\item{BPPARAM}{
    An instance of a \code{BiocParallelParam} class, e.g.,
    \code{\link[BiocParallel]{MulticoreParam}}, \code{\link[BiocParallel]{SnowParam}},
    \code{\link[BiocParallel]{DoparParam}}.
  }
}
\value{
invisible(NULL), files saved to disc
}
\description{
Step 1 of uORFome pipeline: Save leaders and uORF search region
(leader + cds)
}
