% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseGetters.R
\name{getORFNamesDB}
\alias{getORFNamesDB}
\title{Get orf names from the orf data-base}
\usage{
getORFNamesDB(
  with.transcript = F,
  only.transcripts = F,
  asCharacter = T,
  uniques = F
)
}
\arguments{
\item{with.transcript}{a logical(F), should the transcript be included, this makes the
list have duplicated orfs}

\item{only.transcripts}{a logical(F), should only the transcript and not orfId be included}

\item{asCharacter}{a logical(T), should it return as character or data.table(F)}
}
\description{
This is the primary key for most tables in the data-base
}
