% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_analysis.R
\name{getORFsGeneSymbols}
\alias{getORFsGeneSymbols}
\title{Get indices of uORFs from genes}
\usage{
getORFsGeneSymbols(hgncSymbol = "ATF4", refTable = refTable)
}
\arguments{
\item{hgncSymbol}{a character vector of gene symbols}
}
\value{
a integer vector of indices
}
\description{
Get indices of uORFs from genes
}
