% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_analysis.R
\name{getORFsGoTerms}
\alias{getORFsGoTerms}
\title{Get Go terms}
\usage{
getORFsGoTerms(geneNames, organism = "Homo sapiens")
}
\arguments{
\item{geneNames}{ensembl gene names}

\item{organism}{scientifi name}
}
\description{
Get Go terms
}
