% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{getSequenceFeatures}
\alias{getSequenceFeatures}
\title{Get sequence features from orfik}
\usage{
getSequenceFeatures(organism, biomart)
}
\arguments{
\item{organism}{scientifi name, example "Homo sapiens" or "Danio rerio"}

\item{biomart}{name of biomart for organism, example "hsapiens_gene_ensembl"
or "drerio_gene_ensembl"}
}
\description{
Get sequence features from orfik
}
