% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicGetters.R
\name{getSpecialThreeUTRs}
\alias{getSpecialThreeUTRs}
\title{Used as negative set for training
Defined as stop site of 3' UTRs that has width >= 6,
and extend in direction of downstream direction the median
distance of 3' UTRs.}
\usage{
getSpecialThreeUTRs()
}
\description{
Used as negative set for training
Defined as stop site of 3' UTRs that has width >= 6,
and extend in direction of downstream direction the median
distance of 3' UTRs.
}
