% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicGetters.R
\name{getThreeUTRs}
\alias{getThreeUTRs}
\title{Get the 3' sequences from the gtf file}
\usage{
getThreeUTRs(dataFolder = get("dataFolder", envir = .GlobalEnv))
}
\arguments{
\item{dataFolder}{default: get("dataFolder", envir = .GlobalEnv)}
}
\description{
Get the 3' sequences from the gtf file
}
