% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseGetters.R
\name{getUorfsInDb}
\alias{getUorfsInDb}
\title{get the uorfs in the database}
\usage{
getUorfsInDb(withExons = T, withTranscripts = T, uniqueORFs = T)
}
\arguments{
\item{withExons}{should the uorfs be splitted by exons}

\item{withTranscripts}{should the uorfs have transcript information,
warning, this will duplicate some uorfs.}
}
\value{
a GRangesList or data.table, if(F, F)
}
\description{
get the uorfs in the database
}
