% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseHelpers.R
\name{insertTable}
\alias{insertTable}
\title{Delete the database for uORFome}
\usage{
insertTable(
  Matrix,
  tableName,
  appends = FALSE,
  rmOld = FALSE,
  uorfDB = get("uorfDB", envir = .GlobalEnv)
)
}
\arguments{
\item{Matrix}{the data.table / matrix to insert}

\item{tableName}{name of table in sql database}

\item{appends}{rowbind instead of insert}

\item{rmOld}{(TRUE) allow replacing existing table (if TRUE)}

\item{uorfDB}{the defined uORFome sql database, default:
get("uorfDB", envir = .GlobalEnv)}
}
\value{
invisible(NULL)
}
\description{
Will assing to .GlobalEnv
}
