% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicGetters.R
\name{leaderCage}
\alias{leaderCage}
\title{Get maximum spanning CAGE leaders}
\usage{
leaderCage(with.cds = TRUE)
}
\arguments{
\item{with.cds}{logical, default TRUE (add cds upstream)}
}
\description{
Ensures to span all candidate uORFs found
}
