% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseGroupers.R
\name{linkORFsToTx}
\alias{linkORFsToTx}
\title{Assign transcriptnames to orfs, and find rank for each orf}
\usage{
linkORFsToTx(dataFolder)
}
\description{
Given orfs and transcripts, find all transcripts the orfs are within
and name them by this. Also the second orf in
}
