% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierHelpers.R
\name{makeCombinedPrediction}
\alias{makeCombinedPrediction}
\title{Combine classifier and CAGE data, for final prediction table}
\usage{
makeCombinedPrediction(
  tissue,
  dataFolder = get("dataFolder", envir = .GlobalEnv)
)
}
\arguments{
\item{tissue}{tissue to make combined prediction}
}
\value{
data.table with 3 columns. 1. "prediction" (1 or 0), 2. p0 (probability for FALSE),
p1 (probability for TRUE)
}
\description{
Combine classifier and CAGE data, for final prediction table
}
