% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierTrainingData.R
\name{makeTrainingData}
\alias{makeTrainingData}
\title{Get CDS and 3'UTR TrainingData of ribo seq features}
\usage{
makeTrainingData(
  tissues = "combined",
  features = c("countRFP", "disengagementScores", "entropyRFP", "startRegionCoverage",
    "startRegionRelative", "floss", "ioScore", "ORFScores", "fpkmRFP", "RRS", "RSS",
    "startCodonCoverage")
)
}
\arguments{
\item{features}{features to train model on, must exist in database}

\item{tissue}{Tissue to train on, use "combined" if you want all in one,
first run of training it is a ORFik experiment.}
}
\value{
invisible(NULL), saved to disc
}
\description{
Positive set is cds, negative is downstream region of 3' UTRs
}
