% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseGetters.R
\name{matchByTranscript}
\alias{matchByTranscript}
\title{Takes two tables from the database and extracts the rows of toBeMatched
that matches the txNames in referenced.
Both must have a column called txNames}
\usage{
matchByTranscript(toBeMatched, referenced)
}
\value{
the toBeMatched object matched by txNames
}
\description{
Takes two tables from the database and extracts the rows of toBeMatched
that matches the txNames in referenced.
Both must have a column called txNames
}
