% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init.R
\name{orfikDirs}
\alias{orfikDirs}
\title{Init uORFome}
\usage{
orfikDirs(mainPath, df.rfp, df.rna, df.cage, organism, biomart = "not_decided")
}
\arguments{
\item{mainPath}{folder for uORFome to put results}

\item{organism}{scientific name of organism,
like Homo sapiens, Danio rerio, etc.}

\item{biomart}{default "not_decided" will be automaticly detected by
organism name. Set it if you know.}

\item{dataFolder}{ORFik experiment that contains Annotation}

\item{makeDatabase}{default FALSE, set to TRUE if you want
to predict translated uORFs}
}
\description{
Make directoy structure for orf finding
The main Path is ./.. relative to RCode1/ location
}
