% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_analysis.R
\name{ORFsInGeneMetrics}
\alias{ORFsInGeneMetrics}
\title{Check if any uORFs is found in specific gene}
\usage{
ORFsInGeneMetrics(
  hgncSymbol = "ATF4",
  finalCagePred = readTable("finalCAGEuORFPrediction")
)
}
\description{
Check if any uORFs is found in specific gene
}
