% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseValidation.R
\name{pcaCAGEValidation}
\alias{pcaCAGEValidation}
\title{validate features using
seperate them using pca and quantiles}
\usage{
pcaCAGEValidation(uids1, uids2)
}
\description{
validate features using
seperate them using pca and quantiles
}
