% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseValidation.R
\name{predictionVsCageHits}
\alias{predictionVsCageHits}
\title{Plot top tissues}
\usage{
predictionVsCageHits(saveName = "differential_uORF_usage")
}
\arguments{
\item{saveName}{full name of location to save to, will save as png}
}
\description{
Shows max 20 tissues
}
