% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier.R
\name{predictUorfs}
\alias{predictUorfs}
\title{Predict uORFs using random forrest classification}
\usage{
predictUorfs(
  tissues = readTable("experiment_groups")[[1]],
  ip = "localhost",
  port = 54321,
  nthreads_h2o = max(min(45, as.integer(BiocParallel::bpparam()$workers/2)), 1),
  max_mem_size = "200G",
  mode = "uORF"
)
}
\arguments{
\item{tissues}{Groups to train on, default (readTable("experiment_groups")[[1]]),
use "combined" if you want mean of all groups}

\item{ip}{h2o cluster ip, default: "localhost".}

\item{port}{h2o cluster port, default: 54321}

\item{nthreads_h2o}{number of cores for H20. Default:
\code{max(min(45, as.integer(BiocParallel::bpparam()$workers/2)), 1)}}

\item{max_mem_size}{max allowed memory for H20: default ("200G")}

\item{mode}{character, default "uORF"}
}
\description{
Training data
pos set is by CDS (coding sequences)
Neg seg is by trailers (3'UTRs)
}
