% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseGetters.R
\name{removeIDColumns}
\alias{removeIDColumns}
\title{Remove non numberic columns on left side}
\usage{
removeIDColumns(dt)
}
\arguments{
\item{dt}{data.table to remove from}
}
\description{
Remove non numberic columns on left side
}
