% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier.R
\name{sequenceClassifier}
\alias{sequenceClassifier}
\title{Alternative ORF sequence classifier}
\usage{
sequenceClassifier(prediction, tissue)
}
\arguments{
\item{prediction}{data.table of predictions}

\item{tissue}{Tissues to train on, use all if you want all in one}
}
\description{
Alternative ORF sequence classifier
}
