% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierHelpers.R
\name{strongCDS}
\alias{strongCDS}
\title{strongCDS}
\usage{
strongCDS(coverage, fpkm, startCodonCoverage, fiveRegionRelative)
}
\description{
A filter to say if a CDS has strong indication of translation, used as strong positives
}
