% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{subset.ORF}
\alias{subset.ORF}
\title{Subset orfs by filters given}
\usage{
\method{subset}{ORF}(
  orfs,
  minimum.length = 6,
  must.be.mod.3.0 = TRUE,
  startCodons = c("ATG", "CTG", "TTG"),
  stopCodons = c("TAA", "TGA", "TAG"),
  fa
)
}
\arguments{
\item{startCodons}{character vector, default: c("ATG", "CTG", "TTG", "AAG", "AGG")}

\item{stopCodons}{character vector, default: c("TAA", "TGA", "TAG")}

\item{fa}{a FaFile or path to fasta index file}
}
\value{
GRangesList of subset that is valid by filter
}
\description{
Subset orfs by filters given
}
