% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseHelpers.R
\name{tableNotExists}
\alias{tableNotExists}
\title{Does table exist}
\usage{
tableNotExists(name, exact = TRUE, uorfDB = get("uorfDB", envir = .GlobalEnv))
}
\arguments{
\item{name}{Table name}

\item{exact}{(TRUE), if FALSE do "grep" search instead}

\item{uorfDB}{the defined uORFome sql database, default:
get("uorfDB", envir = .GlobalEnv)}
}
\value{
logical (TRUE if not exists)
}
\description{
Does table exist
}
