% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseValidation.R
\name{test.artificial}
\alias{test.artificial}
\title{Test artificial vs original cds}
\usage{
test.artificial(
  artificial,
  output,
  original = artificial,
  minimum.group.size = 30
)
}
\arguments{
\item{artificial}{path to uORFomePipe run of artificial}

\item{output}{path to save plot}

\item{original}{path to uORFomePipe run of original}
}
\value{
data.table of hits per length of artificial
}
\description{
Will see how well the model handles smaller versions of it self
}
