% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseValidation.R
\name{validateAllFeatures}
\alias{validateAllFeatures}
\title{validate features using brain and hela to see that features
 seperate them
 using pca and quantiles}
\usage{
validateAllFeatures()
}
\description{
validate features using brain and hela to see that features
 seperate them
 using pca and quantiles
}
