% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseValidation.R
\name{validateExperiments}
\alias{validateExperiments}
\title{Validate uorfs of data-base}
\usage{
validateExperiments(grl)
}
\description{
This is a check to see that pipeline have done everything correctly
if redoing the findOverlaps does not find all orfs within fiveUTRs
it means that some orfs are outside the mapping area
this should not happen!
}
