% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init.R
\name{validateInputs}
\alias{validateInputs}
\title{Validation of experiments}
\usage{
validateInputs(df.rfp, df.rna, df.cage)
}
\arguments{
\item{df.rfp}{ORFik experiment of Ribo-seq}

\item{df.rna}{ORFik experiment of RNA-seq}

\item{df.cage}{ORFik experiment of CAGE}
}
\description{
Validation of experiments
}
