% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseValidation.R
\name{varianceTissueUsage}
\alias{varianceTissueUsage}
\title{Venn diagram between two tissues}
\usage{
varianceTissueUsage(
  tissue1,
  tissue2,
  pred = readTable("tissueAtlasByCageAndPred", with.IDs = FALSE)
)
}
\arguments{
\item{tissue1}{logical TRUE / FALSE}

\item{tissue2}{logical TRUE / FALSE}

\item{pred}{default}
}
\description{
Venn diagram between two tissues
}
