% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseValidation.R
\name{venn.diagram.uORFs}
\alias{venn.diagram.uORFs}
\title{venn diagram of uORFs}
\usage{
venn.diagram.uORFs(
  predictions = readTable("tissueAtlasByCageAndPred"),
  filename = "venn_diagramm_uORFs.png",
  width = 5,
  height = 5
)
}
\arguments{
\item{predictions}{a data.table of predictions, default: readTable("tissueAtlasByCageAndPred")}

\item{filename}{'venn_diagramm_uORFs.png'}

\item{width}{numeric, default: 5 (unit is inch)}

\item{height}{numeric, default: 5 (unit is inch)}
}
\value{
invisible(NULL), object saved to disc
}
\description{
Shows overlap between all stages / tissues
}
