% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseValidation.R
\name{verification.conf.matrix}
\alias{verification.conf.matrix}
\title{Get comparison of validation model to full CDS length}
\usage{
verification.conf.matrix(
  artificial,
  output,
  original = artificial,
  minimum.group.size = 30
)
}
\arguments{
\item{artificial}{path to uORFomePipe run of artificial}

\item{output}{path to save plot}

\item{original}{path to uORFomePipe run of original}
}
\description{
Get comparison of validation model to full CDS length
}
