% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseGroupers.R
\name{allLeadersSpanningLeader}
\alias{allLeadersSpanningLeader}
\title{Make a leader that spans the max leader per transcript by cage files}
\usage{
allLeadersSpanningLeader(leadersFolder, dataFolder)
}
\arguments{
\item{leadersFolder}{folder with CAGE leaders}

\item{dataFolder}{dataFolder with helper objects}
}
\value{
return(invisible(NULL))
}
\description{
Make a leader that spans the max leader per transcript by cage files
}
