% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseHelpers.R
\name{deleteDataBase}
\alias{deleteDataBase}
\title{Delete the database for uORFome}
\usage{
deleteDataBase(name, uorfDB = get("uorfDB", envir = .GlobalEnv))
}
\arguments{
\item{name}{full path name of database, existing or fresh.}
}
\description{
Will assing to .GlobalEnv
}
