% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierHelpers.R
\name{getAllUORFData}
\alias{getAllUORFData}
\title{Get all data for combined dataset}
\usage{
getAllUORFData(
  prediction = readTable("finalPredWithProb"),
  stringsAsFactors = TRUE
)
}
\arguments{
\item{prediction}{Which prediction to use, default:
prediction = readTable("finalPredWithProb"}

\item{stringsAsFactors}{TRUE}
}
\value{
data.table of prediction, NGS features and sequence features
}
\description{
Get all data for combined dataset
}
