% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierTrainingData.R
\name{getBestIsoformStartCodonCoverage}
\alias{getBestIsoformStartCodonCoverage}
\title{A filter per stop codon group}
\usage{
getBestIsoformStartCodonCoverage(tissue, cdsAndThree = FALSE)
}
\description{
Uses the ribo-seq libraries for ribo-seq validation
get start for each in group, count overlaps, return orf with
highest per group
}
