% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_analysis.R
\name{getBiomartFromOrganism}
\alias{getBiomartFromOrganism}
\title{Guess biomart from organism name}
\usage{
getBiomartFromOrganism(organism, biomart = "ensembl")
}
\arguments{
\item{organism}{scientific name}

\item{biomart}{character, default "ensembl"}
}
\value{
a character with dataset used
}
\description{
Guess biomart from organism name
}
