% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineParts.R
\name{getCandidateuORFs}
\alias{getCandidateuORFs}
\title{Find uORFs from new leader regions}
\usage{
getCandidateuORFs(
  folder = regionUORFsFolder,
  startCodons = "ATG|CTG|TTG|GTG|AAG|AGG|ACG|ATC|ATA|ATT",
  BPPARAM = bpparam()
)
}
\arguments{
\item{startCodons}{default "ATG|CTG|TTG|GTG|AAG|AGG|ACG|ATC|ATA|ATT"}
}
\description{
Step 2 of uORFome pipeline
}
