% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{getGeneralRiboFeatures}
\alias{getGeneralRiboFeatures}
\title{Get Ribo-seq features}
\usage{
getGeneralRiboFeatures(
  df.rfp,
  df.rna = NULL,
  grl,
  preName = "",
  threeUTRsSpecial = NULL,
  BPPARAM = bpparam()
)
}
\arguments{
\item{df.rfp}{ORFik experiment of Ribo-seq}

\item{df.rna}{NULL}

\item{grl}{the ORFs as GRangesList}

\item{preName}{name to pre append in database for each feature}

\item{threeUTRsSpecial}{default NULL, or a GRangesList of 3' UTRs}

\item{BPPARAM}{
    An instance of a \code{BiocParallelParam} class, e.g.,
    \code{\link[BiocParallel]{MulticoreParam}}, \code{\link[BiocParallel]{SnowParam}},
    \code{\link[BiocParallel]{DoparParam}}.
  }
}
\value{
NULL (features saved to database)
}
\description{
Excluding features that uses RNA-seq normalizations if df.rna is NULL
}
