% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicGetters.R
\name{getGTF}
\alias{getGTF}
\title{Get the Genomic transcript format, currently using GRch38 data}
\usage{
getGTF(assignIt = TRUE)
}
\arguments{
\item{assignIt}{TRUE (assign to .GlobalEnv)}
}
\description{
Get the Genomic transcript format, currently using GRch38 data
}
