% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineParts.R
\name{getLeadersFromCage}
\alias{getLeadersFromCage}
\title{Reassign leaders by CAGE}
\usage{
getLeadersFromCage(cageFiles, BPPARAM = bpparam())
}
\arguments{
\item{cageFiles}{a ORFik experiment with CAGE files and annotation}
}
\value{
uORFs search region (CAGE leaders + CDS)
}
\description{
Step 1 of uORFome pipeline
}
