% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_analysis.R
\name{getORFsGoTerms}
\alias{getORFsGoTerms}
\title{Get Go terms}
\usage{
getORFsGoTerms(geneNames, organism)
}
\arguments{
\item{geneNames}{ensembl gene names}

\item{organism}{scientific name}
}
\value{
a data.table of geneNames and go terms(2 columns)
}
\description{
Get Go terms
}
