% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{getSequenceFeatures}
\alias{getSequenceFeatures}
\title{Get sequence features from orfik}
\usage{
getSequenceFeatures(
  organism,
  biomart,
  grl = uORFomePipe:::getUorfsInDb(),
  mode = "uORF"
)
}
\arguments{
\item{organism}{scientific name of organism,
like Homo sapiens, Danio rerio, etc.}

\item{biomart}{default "ensembl", get gene symbols and GO terms for uORF genes. Will
be automaticly detected by organism name in ensembl database.
Set to NULL if you don't want to check Gene symbols and GO terms.}

\item{grl}{the orfs as GRangesList, default uORFomePipe:::getUorfsInDb()}

\item{mode}{character, default: "uORF". alternative "aCDS". Do you want to predict
on uORFs or artificial CDS. if "aCDS" will run twice once for whole length CDS and one for
truncated CDS to validate model works for short ORFs. "CDS" is option to predict on
whole CDS.}
}
\value{
return(invisible(NULL)
}
\description{
Get sequence features from orfik
}
