% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierTrainingData.R
\name{makeORFPredictionData}
\alias{makeORFPredictionData}
\title{Predict table for uORFs}
\usage{
makeORFPredictionData(
  tissues = "combined",
  mode = "uORF",
  features = c("countRFP", "disengagementScores", "entropyRFP", "floss", "fpkmRFP",
    "ioScore", "ORFScores", "RRS", "RSS", "startCodonCoverage", "startRegionCoverage",
    "startRegionRelative")
)
}
\arguments{
\item{tissues}{Tissue to train on, use "combined" if you want all in one,
first run of training it is a ORFik experiment.}

\item{mode}{character, default: "uORF". alternative "aCDS". Do you want to predict
on uORFs or artificial CDS. if "aCDS" will run twice once for whole length CDS and one for
truncated CDS to validate model works for short ORFs. "CDS" is option to predict on
whole CDS.}

\item{features}{features to train model on, any of the features created
during ORFik::computeFeatures, default:
\code{c("countRFP", "disengagementScores", "entropyRFP", "floss",
"fpkmRFP","ioScore", "ORFScores", "RRS", "RSS", "startCodonCoverage",
"startRegionCoverage","startRegionRelative")}}
}
\description{
Get ribo-seq features for uORFs
Will only load if it already exists
}
