% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierTrainingData.R
\name{makeORFPredictionData}
\alias{makeORFPredictionData}
\title{Predict table for uORFs}
\usage{
makeORFPredictionData(
  tissues = "combined",
  features = c("countRFP", "disengagementScores", "entropyRFP", "startRegionCoverage",
    "startRegionRelative", "floss", "ioScore", "ORFScores", "fpkmRFP", "RRS", "RSS",
    "startCodonCoverage")
)
}
\arguments{
\item{features}{features to train model on, must exist in database}

\item{tissue}{Tissue to train on, use "combined" if you want all in one}
}
\description{
Get ribo-seq features for uORFs
Will only load if it already exists
}
