% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierTrainingData.R
\name{makeORFPredictionData}
\alias{makeORFPredictionData}
\title{Predict table for uORFs}
\usage{
makeORFPredictionData(
  tissue = "all",
  features = c("countRFP", "disengagementScores", "entropyRFP", "startRegionCoverage",
    "startRegionRelative", "floss", "ioScore", "ORFScores", "fpkmRFP", "RRS", "RSS",
    "startCodonCoverage")
)
}
\arguments{
\item{tissue}{default "all"}
}
\description{
Get ribo-seq features for uORFs
Will only load if it already exists
}
