% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineParts.R
\name{makeTrainingAndPredictionData}
\alias{makeTrainingAndPredictionData}
\title{All features from sequence, Riboseq and RNAseq}
\usage{
makeTrainingAndPredictionData(
  df.rfp,
  df.rna,
  tissue,
  organism,
  biomart = get("biomart_dataset", mode = "character", envir = .GlobalEnv)
)
}
\description{
Step 5 of uORFome pipeline
}
