% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init.R
\name{orfikDirs}
\alias{orfikDirs}
\title{Init uORFome}
\usage{
orfikDirs(mainPath, df.rfp, df.rna, df.cage, force.remake = FALSE)
}
\arguments{
\item{mainPath}{folder for uORFome to put results}

\item{force.remake}{logical (default: FALSE), force all folders to be remade.
Will not delete existing folders}

\item{dataFolder}{ORFik experiment that contains Annotation}

\item{makeDatabase}{default FALSE, set to TRUE if you want
to predict translated uORFs}
}
\description{
Make directoy structure for orf finding
The main Path is ./.. relative to RCode1/ location
}
