% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init.R
\name{orfikDirs}
\alias{orfikDirs}
\title{Init uORFome pipeline}
\usage{
orfikDirs(
  mainPath,
  df.rfp,
  df.rna,
  df.cage,
  organism,
  biomart = "ensembl",
  mode = "uORF",
  startCodons.cds.allowed = c("ATG", "CTG", "TTG", "AAG", "AGG"),
  stopCodons.cds.allowed = c("TAA", "TGA", "TAG")
)
}
\arguments{
\item{mainPath}{folder for uORFome to put results}

\item{df.rfp}{ORFik experiment of Ribo-seq}

\item{df.rna}{ORFik experiment of RNA-seq}

\item{df.cage}{ORFik experiment of CAGE, set to NULL if you don't have CAGE.}

\item{organism}{scientific name of organism,
like Homo sapiens, Danio rerio, etc.}

\item{biomart}{default "ensembl", get gene symbols and GO terms for uORF genes. Will
be automaticly detected by organism name in ensembl database.
Set to NULL if you don't want to check Gene symbols and GO terms.}

\item{mode}{character, default: "uORF". alternative "aCDS". Do you want to predict
on uORFs or artificial CDS. if "aCDS" will run twice once for whole length CDS and one for
truncated CDS to validate model works for short ORFs. "CDS" is option to predict on
whole CDS.}

\item{startCodons.cds.allowed}{character, default same as startCodons argument.
Which start codons can the CDS you train on have ?}

\item{stopCodons.cds.allowed}{character, default same as stopCodons argument
Which stop codons can the CDS you train on have ?}
}
\description{
Make directory structure for orf finding, create database
assign variables and validate input data.
}
\details{
Debug possibility: \cr
If you get this error: \cr
 Error in bmRequest(request = request, verbose = verbose) :\cr
 Internal Server Error (HTTP 500). \cr
Rerun and it should work, this is a port issue.
}
