% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseValidation.R
\name{perSampleCountPlot}
\alias{perSampleCountPlot}
\title{Per sample library plot overall counts}
\usage{
perSampleCountPlot(mode = "uORF")
}
\arguments{
\item{mode}{character, default: "uORF". alternative "ORF" or what ever
region you are checking.}
}
\description{
Divide into active and candidate uORFs
}
