% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseValidation.R
\name{predictionVsCageHits}
\alias{predictionVsCageHits}
\title{Plot top tissues}
\usage{
predictionVsCageHits(
  saveName = p("differential_", mode, "_usage.png"),
  mode = "uORF"
)
}
\arguments{
\item{saveName}{full name of location to save to, default
"differential_uORF_usage.png"}

\item{mode}{character, default: "uORF". alternative "ORF" or what ever
region you are checking.}
}
\description{
Shows max 20 tissues candidate and predicted uORF counts
And start and stop codon usage for combined prediction
}
