% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseValidation.R
\name{predictionVsCageHits}
\alias{predictionVsCageHits}
\title{Plot top tissues}
\usage{
predictionVsCageHits(
  saveName = p("differential_", mode, "_usage.png"),
  mode = "uORF",
  prediction = predictUorfs(mode = mode),
  stop.codons = FALSE,
  sort.by = "counts"
)
}
\arguments{
\item{saveName}{full name of location to save to, default
"differential_uORF_usage.png"}

\item{prediction}{data.table of prediction values in c(0, 1).}

\item{stop.codons}{= FALSE, if TRUE add a row of stop codon plots}
}
\description{
Shows max 20 tissues candidate and predicted uORF counts
And start and stop codon usage for combined prediction
}
