% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseHelpers.R
\name{readTable}
\alias{readTable}
\title{Read a table from the database}
\usage{
readTable(
  tableName,
  asGR = FALSE,
  with.IDs = TRUE,
  uorfDB = get("uorfDB", envir = .GlobalEnv)
)
}
\arguments{
\item{tableName}{name of table in sql database}

\item{asGR}{convert to GRanges}

\item{with.IDs}{include ID column (remove to make calculations easier)}
}
\value{
the table as data.table or GRanges
}
\description{
Read a table from the database
}
