% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBaseValidation.R
\name{startAndStopCodonPlots}
\alias{startAndStopCodonPlots}
\title{Distribution of start and stop codon according to total prediction}
\usage{
startAndStopCodonPlots()
}
\description{
Distribution of start and stop codon according to total prediction
}
