% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifierStats.R
\name{startCodonMetrics}
\alias{startCodonMetrics}
\title{Get start codon bias usage}
\usage{
startCodonMetrics(
  hits,
  tissue = NULL,
  StartCodons = getAllSequenceFeaturesTable()$StartCodons
)
}
\arguments{
\item{hits}{a logical vector of TRUE, FALSE (TRUE is predicted)}

\item{tissue}{default NULL, or a character name of tissue to print}

\item{StartCodons}{character, sequences of start codons, default:
getAllSequenceFeaturesTable()$StartCodons}
}
\description{
Usually something like this
good: ATG, CTG, ACG and GTG
bad: AAG AND AGG
}
