% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesHelpers.R
\name{toGRFromUniqueID}
\alias{toGRFromUniqueID}
\title{Create GRangesList of ORF IDs}
\usage{
toGRFromUniqueID(uniqueIDs)
}
\arguments{
\item{uniqueIDs}{a character vector of ORF ids, or single column data.table}
}
\value{
GRangesList
}
\description{
Create GRangesList of ORF IDs
}
