% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier.R
\name{trainClassifier}
\alias{trainClassifier}
\title{The training model with cds and 3' UTRs as random forest}
\usage{
trainClassifier(tissue = NULL, nthreads, max_mem_size)
}
\arguments{
\item{tissue}{Tissues to train on, use all if you want all in one}
}
\description{
The training model with cds and 3' UTRs as random forest
}
