% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init.R
\name{validateInputs}
\alias{validateInputs}
\title{Validation of experiments}
\usage{
validateInputs(df.rfp, df.rna, df.cage, mode)
}
\arguments{
\item{df.rfp}{ORFik experiment of Ribo-seq}

\item{df.rna}{ORFik experiment of RNA-seq,
set to NULL if you don't have RNA-seq}

\item{df.cage}{ORFik experiment of CAGE,
set to NULL if you don't have CAGE.}

\item{mode}{character, default: "uORF". alternative "aCDS". Do you want to predict
on uORFs or artificial CDS. if "aCDS" will run twice once for whole length CDS and one for
truncated CDS to validate model works for short ORFs. "CDS" is option to predict on
whole CDS.}
}
\description{
Validation of experiments
}
